+++ src/process/process/gcl_farm/gcl_farm_manager_process.ts
@@ -1,7 +1,7 @@
 import { Procedural } from "process/procedural"
 import { Process, ProcessId } from "process/process"
 import { RoomName } from "utility/room_name"
-import { coloredText, profileLink, roomLink } from "utility/log"
+import { coloredText, roomLink } from "utility/log"
 import { ProcessState } from "../../process_state"
 import { ProcessDecoder } from "../../process_decoder"
 import { GclFarmProcess } from "./gcl_farm_process"
@@ -10,30 +10,12 @@ import { RoomResources } from "room_resource/room_resources"
 import { PrimitiveLogger } from "os/infrastructure/primitive_logger"
 import { MessageObserver } from "os/infrastructure/message_observer"
 import { ListArguments } from "os/infrastructure/console_command/utility/list_argument_parser"
-import { Invader } from "game/invader"
-import { GameConstants } from "utility/constants"
-import { Timestamp } from "utility/timestamp"
-import { Position } from "prototype/room_position"
+import { GclFarmRoom } from "./gcl_farm_types"
+import { validateGclFarmTarget } from "./gcl_farm_target_validator"
 
 ProcessDecoder.register("GclFarmManagerProcess", state => {
   return GclFarmManagerProcess.decode(state as GclFarmManagerProcessState)
 })
-
-type GclFarmRoom = {
-  readonly roomName: RoomName
-  readonly parentRoomNames: RoomName[]
-}
-
-type GclFarmPlan = {
-  storagePosition: Position,
-  terminalPosition: Position,
-  spawnPosition: Position,
-  distributorPosition: Position,
-  tower1Position: Position,
-  tower2Position: Position,
-  tower3Position: Position,
-  upgraderPositions: Position[],
-}
 
 interface GclFarmManagerProcessState extends ProcessState {
   readonly name: string
@@ -156,7 +138,7 @@ export class GclFarmManagerProcess implements Process, Procedural, MessageObserv
       parentRoomNames,
     }
 
-    const validationErrors = this.validateTarget(target)
+    const validationErrors = validateGclFarmTarget(target)
     if (validationErrors.errors.length > 0) {
       throw `Invalid target ${roomLink(targetRoomName)}\n${validationErrors.errors.join("\n")}`
     }
@@ -240,7 +222,7 @@ export class GclFarmManagerProcess implements Process, Procedural, MessageObserv
       return
     }
 
-    const validationErrors = this.validateTarget(target)
+    const validationErrors = validateGclFarmTarget(target)
     if (validationErrors.errors.length > 0) {
       PrimitiveLogger.fatal(`${this.constructor.name} ${this.processId} cannot launch GCL farm on ${roomLink(target.roomName)}\n${validationErrors.errors.join("\n")}`)
       OperatingSystem.os.suspendProcess(this.processId)
@@ -251,84 +233,5 @@ export class GclFarmManagerProcess implements Process, Procedural, MessageObserv
     //   return GclFarmProcess.create(processId, target.roomName, target.parentRoomNames)
     // })
 
-  }
-
-  /**
-   * - 条件：
-   *   - targetはInvader以外にclaimされていない
-   *   - targetはInvaderか自分以外にreserveされていない
-   *   - parent roomsはtargetの隣（通路が空いている）
-   *   - parent roomsはRCL8以上
-   *   - parent roomsは完全（Spawn, Storage, Terminal等の存在
-   */
-  private validateTarget(target: GclFarmRoom): { errors: string[] } {
-    const errors: string[] = []
-    const targetRoomInfo = RoomResources.getRoomInfo(target.roomName)
-
-    if (targetRoomInfo == null) {
-      return {
-        errors: [`target ${roomLink(target.roomName)} is not observed`]
-      }
-    }
-
-    const getRelativeTime = (timestamp: Timestamp): string => {
-      if (timestamp < 1000) {
-        return `${Game.time - timestamp} ticks ago`
-      }
-      return `${Math.floor((Game.time - timestamp) / 1000)}k ticks ago`
-    }
-
-    switch (targetRoomInfo.roomType) {
-    case "owned":
-      errors.push(`target ${roomLink(target.roomName)} is already mine`)
-      break
-    case "normal":
-      if (targetRoomInfo.owner != null) {
-        switch (targetRoomInfo.owner.ownerType) {
-        case "claim":
-          if (targetRoomInfo.owner.username !== Invader.username) {
-            errors.push(`target ${roomLink(target.roomName)} is owned by ${profileLink(targetRoomInfo.owner.username)} ${getRelativeTime(targetRoomInfo.observedAt)}`)
-          }
-          break
-        case "reserve":
-          if ([Invader.username, Game.user.name].includes(targetRoomInfo.owner.username) !== true) {
-            errors.push(`target ${roomLink(target.roomName)} is reserved by ${profileLink(targetRoomInfo.owner.username)} ${getRelativeTime(targetRoomInfo.observedAt)}`)
-          }
-          break
-        }
-      }
-      break
-    }
-
-    const neighbourRoomNames = [...targetRoomInfo.neighbourRoomNames]
-
-    target.parentRoomNames.forEach(parentRoomName => {
-      if (neighbourRoomNames.includes(parentRoomName) !== true) {
-        errors.push(`parent room ${roomLink(parentRoomName)} is not next to the target ${roomLink(target.roomName)}`)
-        return
-      }
-
-      const parentRoomResource = RoomResources.getOwnedRoomResource(parentRoomName)
-      if (parentRoomResource == null) {
-        errors.push(`parent room ${roomLink(parentRoomName)} is not owned`)
-        return
-      }
-      if (parentRoomResource.controller.level < 8) {
-        errors.push(`parent room ${roomLink(parentRoomName)} is under development (RCL${parentRoomResource.controller.level})`)
-        return
-      }
-      if (
-        parentRoomResource.activeStructures.terminal == null
-        || parentRoomResource.activeStructures.storage == null
-        || parentRoomResource.activeStructures.spawns.length < GameConstants.structure.maxCount.spawn
-      ) {
-        errors.push(`parent room ${roomLink(parentRoomName)} lack of vital structures`)
-        return
-      }
-    })
-
-    return {
-      errors,
-    }
   }
 }
